refact: View를 입력/출력으로 분리
