+++ src/main/java/racingcar/view/UserOutputView.java
@@ -4,27 +4,12 @@ import racingcar.domain.Car;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Scanner;
 
-public class UserView {
+public class UserOutputView {
 
-    private static final String REQUEST_CAR_NAMES = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
-    private static final String REQUEST_ROUND = "시도할 회수는 몇회인가요?";
     private static final String PROCESS_TITLE = "실행 결과";
     private static final String WINNER_MESSAGE = "(이)가 최종 우승했습니다.";
     private static final String STATUS_BAR_STYLE = "-";
-
-    public static String enterCarNames() {
-        System.out.println(REQUEST_CAR_NAMES);
-        Scanner scanner = new Scanner(System.in);
-        return scanner.nextLine();
-    }
-
-    public static int enterRound() {
-        System.out.println(REQUEST_ROUND);
-        Scanner scanner = new Scanner(System.in);
-        return scanner.nextInt();
-    }
 
     public static void printProcessTitle() {
         System.out.println(PROCESS_TITLE);
